requirements.txt
Django>=5.2.4
djangorestframework>=3.14.0
django-filter>=23.2
django-cors-headers>=4.0.0
requests>=2.31.0
drf-spectacular>=0.29.0

